+++ example/lib/widgets/record_list_item.dart
@@ -22,8 +22,13 @@ class RecordListItem extends StatelessWidget {
       final BloodOxygenSample r => '${r.percentage.toStringAsFixed(1)}%',
       final StressScore r => 'Score: ${r.score}',
       final SleepScore r => 'Score: ${r.score}',
-      final HeartRateVariability r =>
-        '${r.sdnnMilliseconds.toStringAsFixed(1)} ms',
+      final HeartRateVariability r => switch (r) {
+          HeartRateVariability(:final sdnnMilliseconds?) =>
+            'SDNN ${sdnnMilliseconds.toStringAsFixed(1)} ms',
+          HeartRateVariability(:final rmssdMilliseconds?) =>
+            'RMSSD ${rmssdMilliseconds.toStringAsFixed(1)} ms',
+          _ => '-- ms',
+        },
       final RestingHeartRate r => '${r.beatsPerMinute} bpm',
       final RespiratoryRate r =>
         '${r.breathsPerMinute.toStringAsFixed(1)} br/min',